My attempt at Dump function
Thanks again for writing this package! I am pretty new to Golang, but I really like this
language, so I have been trying to use it as my main language. Feel free on any ideas you have
to make this pull request better.

Side note: I made the second param into Dump an io.Writer because it made the functionality
easier to test and it did not limit the functionality. Let me know your thoughts on it