Made the changes
I made a small wrapper function that writes the data to a file (DumpToFile).
I added a test to check the error case where the an empty string is passed in.
I removed the Dump line from the Todo section of the README.

I thought about adding and example of using the DumpToFile function, but
I was not sure where to put it or what kind of example I should make