My attempt at Dump function (#6)
* My attempt at Dump function

Thanks again for writing this package! I am pretty new to Golang, but I really like this
language, so I have been trying to use it as my main language. Feel free on any ideas you have
to make this pull request better.

Side note: I made the second param into Dump an io.Writer because it made the functionality
easier to test and it did not limit the functionality. Let me know your thoughts on it

* Made the changes

I made a small wrapper function that writes the data to a file (DumpToFile).
I added a test to check the error case where the an empty string is passed in.
I removed the Dump line from the Todo section of the README.

I thought about adding and example of using the DumpToFile function, but
I was not sure where to put it or what kind of example I should make

* Add Dump example

* Add README example to test cases